# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AraLink API",
    "version": "1.0.0",
    "description": "AraLink Translation API — ترجم أي رابط (يوتيوب، مقالات، مواقع) إلى أي لغة.\n\nجميع نقاط النهاية سلمت تحت `/api`. نقاط الإدارة تتطلب ترويسة `x-admin-token` (انظر ADMIN_TOKEN في .env.example)."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Development" }
  ],
...
  "components": {
    "securitySchemes": {
      "AdminToken": {
        "type": "apiKey",
        "in": "header",
        "name": "x-admin-token",
        "description": "رمز الإدارة ADMIN_TOKEN — يفتح مسارات /api/settings و /api/stats وإحصاءات /api/jobs"
      }
    },
    "schemas": {
      "TranslationRequest": {
        "type": "object",
        "required": ["url", "targetLang"],
        "properties": {
          "url": { "type": "string", "format": "uri", "description": "رابط يوتيوب أو مقال أو موقع" },
          "targetLang": { "type": "string", "examples": ["ar"], "description": "رمز اللغة الهدف (مثل ar أو en أو de)" },
          "sourceLang": { "type": "string", "description": "رمز اللغة المصدر — يُكتشف تلقائيًا عند الإهمال" },
          "provider": { "type": "string", "description": "فرض مزوّد واحد (مثل google). عند فشله لا يوجد احتياط" },
          "providers": { "type": "array", "items": { "type": "string" }, "description": "سلسلة مزوّدين مرتبة تُجرَّب بالترتيب" }
        }
      },
...
      "TranslationResponse": {
        "type": "object",
        "properties": {
          "type": { "type": "string", "examples": ["youtube", "article", "text"] },
          "sourceLang": { "type": "string", "description": "اللغة المكتشفة أو الممرَّرة" },
          "translated": { "type": "string", "description": "النص المترجم — الأسطر تطابق الأصل سطرًا بسطر" },
          "original": { "type": "string" },
          "meta": { "type": "object", "additionalProperties": true, "description": "عنوان/مدة/مزوّد... حسب النوع" }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "رمز خطأ قصير (مثل video-not-found)" },
          "message": { "type": "string", "description": "وصف عربي قابل للعرض" },
          "code": { "type": "string" }
        }
      },
...
      "JobStatus": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "status": { "type": "string", "enum": ["queued", "running", "done", "failed", "cancelled"] },
          "progress": { "type": "number", "minimum": 0, "maximum": 100 },
          "result": { "type": "object", "additionalProperties": true },
          "error": { "type": "string" }
        }
      }
    },
    "parameters": {
 "AdminHeader": {
   "name": "x-admin-token",
   "in": "header",
   "required": true,
   "schema": {
    "type": "string"
   }
  },
...
    }
  },
  "paths": {
    "/api/admin/login": {
      "post": {
        "tags": ["admin"],
        "summary": "تبديل رمز الإدارة بكوكي httpOnly",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["token"],
                "properties": { "token": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "تم الدخول — كوكي httpOnly" },
          "401": { "description": "رمز خاطئ", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "503": { "description": "ADMIN_TOKEN غير مضبوط", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/api/translate": {
      "post": {
        "tags": ["public"],
        "summary": "ترجمة رابط (يوتيوب/مقال/موقع)",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TranslationRequest" } } }
        },
        "responses": {
          "200": {
            "description": "الترجمة نجحت",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TranslationResponse" } } }
          },
          "400": { "description": "طلب ناقص أو رابط غير صالح", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "الفيديو أو الصفحة غير موجودة", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "429": { "description": "تجاوز حد الطلبات", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "502": { "description": "فشل جلب المحتوى من المصدر", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/api/translate-smart": {
      "post": {
        "tags": ["public"],
        "summary": "ترجمة ذكية: نص أو رابط — يكتشف النوع تلقائيًا",
        "description": "يقبل نصًا حرًا أو رابطًا في حقل واحد ويوجه للمسار المناسب داخليًا.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["input"],
                "properties": {
                  "input": { "type": "string", "description": "نص حر أو رابط" },
                  "targetLang": { "type": "string" },
                  "sourceLang": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "الترجمة نجحت", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TranslationResponse" } } } },
          "400": { "description": "مدخل ناقص", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/api/translate-file": {
      "post": {
        "tags": ["public"],
        "summary": "ترجمة ملف (docx/pdf/epub/srt/md/txt...) وإرجاعه بصيغة قابلة للتنزيل",
        "description": "حد الجسم 15MB. الصيغ المدعومة تديرها server/files.js.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["content", "format", "targetLang"],
                "properties": {
                  "content": { "type": "string", "description": "محتوى الملف (نصي أو base64 حسب الصيغة)" },
                  "format": { "type": "string", "examples": ["txt", "srt", "md"] },
                  "targetLang": { "type": "string" },
                  "sourceLang": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "الملف المترجم", "content": { "application/octet-stream": { "schema": { "type": "string", "format": "binary" } } } },
          "400": { "description": "صيغة غير مدعومة أو طلب ناقص", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/export": {
      "post": {
        "tags": ["public"],
        "summary": "تصدير نص مترجم إلى docx/pdf/xlsx/md/txt",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["content", "format"],
                "properties": {
                  "content": { "type": "string" },
                  "format": { "type": "string", "examples": ["docx", "pdf", "xlsx", "md", "txt"] },
                  "filename": { "type": "string", "description": "اسم مقترح للتنزيل — يُستبدل باسم آمن دائمًا" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "ملف التصدير", "content": { "application/octet-stream": { "schema": { "type": "string", "format": "binary" } } } },
          "400": { "description": "صيغة غير مدعومة", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/ocr": {
      "post": {
        "tags": ["media"],
        "summary": "استخراج نص من صورة عبر OCR (Tesseract)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["image"],
                "properties": {
                  "image": { "type": "string", "format": "byte", "description": "الصورة base64" },
                  "lang": { "type": "string", "examples": ["ara", "eng"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "النص المستخرج", "content": { "application/json": { "schema": { "type": "object", "properties": { "text": { "type": "string" } } } } } },
          "400": { "description": "صورة غير صالحة", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/tts": {
      "post": {
        "tags": ["media"],
        "summary": "تحويل نص إلى صوت (Edge voices أو gTTS احتياطيًا)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text", "lang"],
                "properties": {
                  "text": { "type": "string", "maxLength": 5000 },
                  "lang": { "type": "string", "examples": ["ar"] },
                  "voice": { "type": "string", "description": "اسم صوت Edge اختياري" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "ملف صوتي", "content": { "audio/mpeg": { "schema": { "type": "string", "format": "binary" } } } },
          "400": { "description": "نص أو لغة ناقصة", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/tashkeel": {
      "post": {
        "tags": ["public"],
        "summary": "إضافة التشكيل للنص العربي",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["text"],
                "properties": { "text": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "النص مشكّلًا", "content": { "application/json": { "schema": { "type": "object", "properties": { "text": { "type": "string" } } } } } },
          "400": { "description": "نص ناقص", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/video-local": {
      "post": {
        "tags": ["media"],
        "summary": "ترجمة فيديو مرفوع محليًا (تفريغ ثم ترجمة)",
        "description": "حد الجسم 60MB ومدة أقصى افتراضية 5 دقائق (LOCAL_VIDEO_MAX_MIN). يعيد كائن مهمة — تابع التقدم عبر /api/jobs/{id}.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["video"],
                "properties": {
                  "video": { "type": "string", "format": "byte", "description": "ملف الفيديو base64" },
                  "targetLang": { "type": "string" },
                  "sourceLang": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "مهمة بدأت", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobStatus" } } } },
          "400": { "description": "ملف غير صالح أو تجاوز الحد", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "413": { "description": "الملف أكبر من الحد", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/api/youtube/dub": {
      "post": {
        "tags": ["media"],
        "summary": "دبلجة فيديو يوتيوب: تنزيل ثم تفريغ ثم ترجمة ثم إعادة نطق",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url", "targetLang"],
                "properties": {
                  "url": { "type": "string", "format": "uri" },
                  "targetLang": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "مهمة دبلجة بدأت", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobStatus" } } } },
          "400": { "description": "طلب ناقص", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "429": { "description": "تجاوز حد الدبلجة", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/dub": {
      "post": {
        "tags": ["media"],
        "summary": "دبلجة مقاطع جاهزة (دفعات TTS)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "projectId": { "type": "string" },
                  "segments": { "type": "array", "maxItems": 40, "items": { "type": "object", "additionalProperties": true } }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "المقاطع الصوتية", "content": { "application/json": { "schema": { "type": "object", "additionalProperties": true } } } },
          "400": { "description": "دُفعة غير صالحة", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/api/dub/jobs/{id}/stream": {
      "get": {
        "tags": ["media"],
        "summary": "بث تقدم مهمة دبلجة (SSE)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "تدفق أحداث SSE", "content": { "text/event-stream": { "schema": { "type": "string" } } } },
          "403": { "description": "لست مالك المهمة", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/api/projects": {
      "get": {
        "tags": ["projects"],
        "summary": "سرد كل المشاريع (إدارة فقط)",
        "security": [{ "AdminToken": [] }],
        "responses": {
          "200": { "description": "قائمة المشاريع", "content": { "application/json": { "schema": { "type": "object", "properties": { "projects": { "type": "array", "items": { "type": "object" } } } } } } },
          "401": { "description": "بلا x-admin-token", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "post": {
        "tags": ["projects"],
        "summary": "إنشاء مشروع دبلجة جديد",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "sourceType": { "type": "string", "examples": ["youtube", "upload"] },
                  "sourceRef": { "type": "string" },
                  "targetLangs": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "المشروع أُنشئ (يُعاد id وownerToken لاحقًا في الترويسات/الجسم حسب التنفيذ)" },
          "400": { "description": "اسم ناقص", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/api/jobs/{id}/stream": {
      "get": {
        "tags": ["jobs"],
        "summary": "بث تقدم وظيفة (SSE)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": { "description": "تدفق أحداث SSE", "content": { "text/event-stream": { "schema": { "type": "string" } } } },
          "404": { "description": "غير موجودة", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/settings": {
      "get": {
        "tags": ["admin"],
        "summary": "قراءة الإعدادات الحالية (مفاتيح مقنّعة)",
        "security": [{ "AdminToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AdminHeader" }],
        "responses": {
          "200": { "description": "الإعدادات الحالية" },
          "401": { "description": "رمز إدارة خاطئ", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "503": { "description": "ADMIN_TOKEN غير مضبوط — المسار معطّل", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "post": {
        "tags": ["admin"],
        "summary": "كتابة قيم .env (مفاتيح المزوّدين)",
        "security": [{ "AdminToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AdminHeader" }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": { "type": "string" },
                "description": "خريطة مفتاح→قيمة مثل { GEMINI_API_KEY: '...' }"
              }
            }
          }
        },
        "responses": {
          "200": { "description": "تمت الكتابة" },
          "401": { "description": "رمز إدارة خاطئ", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
    "/api/stats/languages": {
      "get": {
        "tags": ["admin"],
        "summary": "توزيع اللغات المصدر والهدف",
        "security": [{ "AdminToken": [] }],
        "parameters": [{ "$ref": "#/components/parameters/AdminHeader" }],
        "responses": {
          "200": { "description": "{ bySource, byTarget }" },
          "401": { "description": "بلا رمز إدارة", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
  }
}